Track attachments per test method, rather than just per test class.
This means that on the pages for each test method, only any attachments
for that method are shown.  A summary of all attachments for a given
test class are shown on the test class page, along with any attachments
found in the stdout and stderr output of the parent testsuite.

Also, attachments should not be redundantly copied to the master if
they happen to be referenced multiple times within a single testsuite.

This should fix JENKINS-7056 and JENKINS-13115.